Set value of mExtendedAssignmentConsistancyWeight to avoid notebook crash
In an older version of the code, Cell in pysheaf.py had the attribute
mExtendedAssignmentConsistancyWeightDivisor that was defaulted to 2.0. In the
new version, it's replaced with mExtendedAssignmentConsistancyWeight that's
defaulted to None. As a result, running Cell 8 in SheafBuildingExample.ipynb
would lead to a norm being computed on an empty array in Line 181 of pysheaf.py
in ComputeConsistency() and would cause a crash.

This commit assigns mExtendedAssignmentConsistancyWeight=0.5 to
reproduce the original behavior of the notebook.